multiply

+++ 0x0A-argc_argv/3-mul.c
@@ -8,7 +8,7 @@
 
 int main(int argc, char argv[])
 {
-	long product;
+       int product;
 
 	if (argc != 3)
 	{
